% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{check_same_partition}
\alias{check_same_partition}
\title{Check if Two Clusterings Represent Same Partition}
\usage{
check_same_partition(c1, c2)
}
\arguments{
\item{c1}{First clustering vector}

\item{c2}{Second clustering vector}
}
\value{
Logical; TRUE if partitions are identical
}
\description{
Compares two clustering vectors to determine if they represent the same partition.
}
\keyword{internal}
