% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coclustering}
\alias{coclustering}
\title{Compute Co-clustering Probability Matrix}
\usage{
coclustering(samples)
}
\arguments{
\item{samples}{A matrix of cluster assignments where rows are MCMC iterations
and columns are items/observations.}
}
\value{
A symmetric matrix of size n x n (where n is the number of items)
containing the co-clustering probabilities. Entry (i,j) represents the
proportion of MCMC iterations where items i and j were in the same cluster.
}
\description{
Computes the co-clustering probability matrix from MCMC samples,
where each entry (i,j) represents the proportion of iterations
in which items i and j were assigned to the same cluster.
}
