% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tessellation_joint.R
\name{compute_conditional_coupling_logloss}
\alias{compute_conditional_coupling_logloss}
\title{Compute conditional coupling log-loss}
\usage{
compute_conditional_coupling_logloss(curr_view, target_z, z_list, params)
}
\arguments{
\item{curr_view}{Integer; the index of the view currently being updated}

\item{target_z}{Numeric vector; the proposed (or current) tessellation for view \code{curr_view}}

\item{z_list}{List of numeric vectors; the current tessellations for ALL views}

\item{params}{List of model hyperparametes (here uses alpha.a and alpha.b)}
}
\value{
A single numeric value representing the total condition coupling log-loss
}
\description{
Computes the conditional coupling log-loss between the clustering in the current view and the clustering in all other views.
}
\keyword{internal}
