% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_EB_params.R
\name{compute_EB_params}
\alias{compute_EB_params}
\title{Compute Empirical Bayes Parameters}
\usage{
compute_EB_params(
  distance_matrix,
  n_clusters,
  initial_allocs = NULL,
  linear = FALSE,
  repulsion = TRUE
)
}
\arguments{
\item{distance_matrix}{A square numeric matrix of pairwise distances between data points.}

\item{n_clusters}{The number of clusters for the initial k-medoids clustering. This is ignored if `initial_allocs` is provided.}

\item{initial_allocs}{Optional initial cluster assignments. If \code{NULL} (the default), k-medoids clustering is performed using `n_clusters`.}

\item{linear}{Logical. If \code{TRUE}, uses a faster estimation method based on distances to/between medoids. If \code{FALSE} (the default), uses all pairwise distances, which is more accurate but slower.}

\item{repulsion}{Logical. If \code{TRUE} (the default), the returned parameters will indicate that the model should account for between-cluster repulsion.}
}
\value{
A list of likelihood parameters in the format required by \code{mcmc_tessellation} and \code{mcmc_PY}. See \code{\link{likelihood_params}}.
}
\description{
Computes empirical Bayes hyperparameters for the likelihood model from data using k-medoids clustering.
}
