% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_EB_params_linear}
\alias{compute_EB_params_linear}
\title{Linear Empirical Bayes Parameters}
\usage{
compute_EB_params_linear(D, K, k2 = NULL, repulsion = TRUE)
}
\arguments{
\item{D}{Distance matrix}

\item{K}{Number of clusters}

\item{k2}{Optional cluster assignments; if NULL, k-medoids clustering is performed}

\item{repulsion}{Optional parameter: will the sampler consider the repulsive term in the marginal likelihood?}
}
\value{
List of hyperparameters
}
\description{
Computes empirical Bayes hyperparameters using linear model.
}
\keyword{internal}
