% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_EB_params.R
\name{compute_EB_params_quadratic}
\alias{compute_EB_params_quadratic}
\title{Quadratic Empirical Bayes Parameters}
\usage{
compute_EB_params_quadratic(D, K, k2 = NULL, repulsion = TRUE)
}
\arguments{
\item{repulsion}{Logical. If \code{TRUE} (the default), the returned parameters will indicate that the model should account for between-cluster repulsion.}
}
\value{
A list of likelihood parameters.
}
\description{
Computes empirical Bayes hyperparameters using a "quadratic" complexity approach (all pairwise distances).
}
\keyword{internal}
