% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convergence_diagnostics}
\alias{convergence_diagnostics}
\title{MCMC Convergence Diagnostics}
\usage{
convergence_diagnostics(fit)
}
\arguments{
\item{fit}{List of MCMC samples from mcmc_sampler}
}
\value{
Data frame with ESS and Rhat for each parameter
}
\description{
Computes Rhat and ESS diagnostics for MCMC samples.
}
\keyword{internal}
