% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{elements_in_cluster}
\alias{elements_in_cluster}
\title{Helper Function for Cluster Elements}
\usage{
elements_in_cluster(z, k, i)
}
\arguments{
\item{z}{Cluster allocation vector}

\item{k}{Cluster index}

\item{i}{Element index to exclude}
}
\value{
Numeric vector of element indices
}
\description{
Returns indices of elements in cluster k, excluding element i.
}
\keyword{internal}
