% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tessellation_joint.R
\name{f_diff}
\alias{f_diff}
\title{Helper Functions for Rand Index}
\usage{
f_diff(pair)
}
\arguments{
\item{pair}{Numeric vector of length 2 with cluster labels}
}
\value{
Logical value
}
\description{
Check if two elements are in different clusters.
}
\keyword{internal}
