% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{get_log_prob_D}
\alias{get_log_prob_D}
\title{Compute Log Likelihood of Data Given Clustering}
\usage{
get_log_prob_D(D, z, params, rcpp = TRUE, return_L_matrix = FALSE)
}
\arguments{
\item{D}{Distance matrix}

\item{z}{Cluster allocation vector}

\item{params}{List of model parameters}

\item{rcpp}{Logical; use Rcpp implementation}

\item{return_L_matrix}{Logical; if TRUE return full likelihood matrix}
}
\value{
Log likelihood or list with log likelihood and matrix
}
\description{
Computes log probability of distance matrix given cluster allocations.
}
\keyword{internal}
