% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tesselation.R
\name{get_log_prob_D_tesselation}
\alias{get_log_prob_D_tesselation}
\title{Compute Log Likelihood for Tesselation}
\usage{
get_log_prob_D_tesselation(D, z, centers, params)
}
\arguments{
\item{D}{Distance matrix}

\item{z}{Cluster allocation vector}

\item{centers}{Center indices}

\item{params}{Model parameters}
}
\value{
Log likelihood
}
\description{
Computes log probability of distance matrix for Voronoi tesselation.
}
\keyword{internal}
