% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{get_log_prob_partition}
\alias{get_log_prob_partition}
\title{Compute Log Probability of Partition}
\usage{
get_log_prob_partition(z, M, theta, rcpp = TRUE)
}
\arguments{
\item{z}{Cluster allocation vector}

\item{M}{Mass parameter}

\item{theta}{Discount parameter}

\item{rcpp}{Logical; use Rcpp implementation}
}
\value{
Log probability of partition
}
\description{
Computes log probability of a partition under Pitman-Yor process prior.
}
\keyword{internal}
