% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_plots}
\alias{get_plots}
\title{Generate MCMC Diagnostic Plots}
\usage{
get_plots(fit, lags = 50)
}
\arguments{
\item{fit}{List of MCMC samples from mcmc_sampler}

\item{lags}{Number of lags for ACF plots}
}
\description{
Creates trace, histogram, and ACF plots for MCMC parameters.
}
\keyword{internal}
