% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tesselation.R
\name{get_probs_nested}
\alias{get_probs_nested}
\title{Compute Proposal Probabilities for Nested Voronoi}
\usage{
get_probs_nested(
  D,
  gamma,
  z,
  prior_k_mean,
  tempering,
  update_group = NULL,
  z1.ref
)
}
\arguments{
\item{D}{Distance matrix}

\item{gamma}{Binary indicator of center status}

\item{z}{Cluster assignments}

\item{prior_k_mean}{Prior parameter for number of clusters}

\item{tempering}{Tempering parameter}

\item{update_group}{Indices to consider for updates (NULL = all)}

\item{z1.ref}{Reference clustering to maintain partition structure}
}
\value{
Numeric vector of proposal probabilities
}
\description{
Computes probabilities for proposing changes to centers within nested partition structure.
}
\keyword{internal}
