% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tesselation.R
\name{log_prior_gamma2_gamma1_2}
\alias{log_prior_gamma2_gamma1_2}
\title{Log Prior for Compatible Nested Gamma}
\usage{
log_prior_gamma2_gamma1_2(z1.new, centers2.new)
}
\arguments{
\item{z1.new}{Clustering for layer 1}

\item{centers2.new}{Center indices for layer 2}
}
\value{
Log prior probability
}
\description{
Computes log prior probability for nested clustering configuration that respects layer 1 partition.
}
\keyword{internal}
