% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tesselation.R
\name{log_prior_k}
\alias{log_prior_k}
\title{Log Prior for Number of Clusters}
\usage{
log_prior_k(k, N, prior_param)
}
\arguments{
\item{k}{Number of clusters}

\item{N}{Maximum number of clusters (data size)}

\item{prior_param}{Prior parameter(s). If scalar, uses geometric distribution; otherwise uses provided probability vector}
}
\value{
Log prior probability
}
\description{
Computes log prior probability for a given number of clusters.
}
\keyword{internal}
