% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{log_subor}
\alias{log_subor}
\title{Subordinator Function}
\usage{
log_subor(x, m, alpha = 1)
}
\arguments{
\item{x}{Numeric value}

\item{m}{Number of terms}

\item{alpha}{Discount parameter}
}
\value{
Log of subordinator
}
\description{
Computes the log of a subordinator used in PY process.
}
\keyword{internal}
