% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tesselation.R
\name{marginal_prior_predictive_sampler_tesselation}
\alias{marginal_prior_predictive_sampler_tesselation}
\title{Marginal Prior Predictive Sampler for Tesselation}
\usage{
marginal_prior_predictive_sampler_tesselation(n, N.sim = 300, params)
}
\arguments{
\item{n}{Number of observations}

\item{N.sim}{Number of samples to generate (default 300)}

\item{params}{Model parameters}

\item{rcpp}{Logical; use Rcpp implementation (default TRUE, currently unused)}
}
\value{
List of samples with within and between distance distributions
}
\description{
Generates samples from the prior predictive distribution of distances under tesselation model.
}
