% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcmc_multiview_tessellation}
\alias{mcmc_multiview_tessellation}
\title{Run Reversible Jump MCMC for Multi-View Voronoi Tessellation}
\usage{
mcmc_multiview_tessellation(
  distance_matrices,
  likelihood_params,
  prior_params,
  algo_params
)
}
\description{
Run Reversible Jump MCMC for Multi-View Voronoi Tessellation
}
