% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcmc_PY}
\alias{mcmc_PY}
\title{Run Split-Merge MCMC for Pitman-Yor Process Mixture Model}
\usage{
mcmc_PY(distance_matrix, likelihood_params, prior_params, algo_params)
}
\description{
Run Split-Merge MCMC for Pitman-Yor Process Mixture Model
}
