% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcmc_PY_multiview}
\alias{mcmc_PY_multiview}
\title{Run Split-Merge MCMC for Multi-View Pitman-Yor Mixture Model}
\usage{
mcmc_PY_multiview(
  distance_matrices,
  likelihood_params,
  prior_params,
  algo_params
)
}
\arguments{
\item{distance_matrices}{A list of square numeric matrices, one for each view.}

\item{likelihood_params}{A list containing parameters for the likelihood across all views. See \code{\link{multiview_params}} for details.}

\item{prior_params}{A list containing prior parameters for each view. See \code{\link{multiview_params}} for details.}

\item{algo_params}{A list of algorithm parameters. See \code{\link{algorithm_params}} for details on the mixture model parameters.}
}
\value{
A list containing the multi-view MCMC output:
  \itemize{
    \item \code{views}: A list of lists, where each inner list contains the MCMC output for a single view (see `mcmc_PY` return value).
    \item \code{joint_lpdf}: A vector of the joint log-posterior density values (including coupling) for each retained sample.
  }
}
\description{
This function runs a Split-Merge MCMC sampler for a multi-view Pitman-Yor
process mixture model, coupling multiple views.
}
\details{
Run Split-Merge MCMC for Multi-View Pitman-Yor Mixture Model
}
