% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY_marginal.R
\name{mcmc_sampler}
\alias{mcmc_sampler}
\title{MCMC Sampler for Distance Clustering}
\usage{
mcmc_sampler(
  D,
  params,
  N.sim,
  verbose = FALSE,
  rcpp = TRUE,
  M.fixed = NULL,
  theta.fixed = NULL
)
}
\arguments{
\item{D}{Distance or data matrix (n x n matrix)}

\item{params}{List of MCMC parameters including:
\itemize{
\item z.init: Initial cluster allocations
\item M.init: Initial value for M parameter
\item theta.init: Initial value for theta parameter
\item metropolis_M_sd: Proposal standard deviation for M
\item metropolis_theta_sd: Proposal standard deviation for theta
\item lambda.M: Rate parameter for exponential prior on M
\item alpha.theta, beta.theta: Shape parameters for beta prior on theta
\item delta1, mu, beta: Hyperparameters for within-cluster distances
\item delta2, zeta, gamma: Hyperparameters for between-cluster distances
}}

\item{N.sim}{Integer number of MCMC iterations to run}

\item{verbose}{Logical; if TRUE, print iteration details (default FALSE)}

\item{rcpp}{Logical; if TRUE, use Rcpp implementations for speed (default TRUE)}

\item{M.fixed}{Optional fixed value for M parameter; if NULL, M is sampled}

\item{theta.fixed}{Optional fixed value for theta parameter; if NULL, theta is sampled}
}
\value{
List of samples, each containing:
\itemize{
\item M: Sampled M value
\item theta: Sampled theta value
\item z: Cluster allocations
\item K: Number of clusters
\item log_lik: Log-likelihood for WAIC
\item time: Elapsed time
}
}
\description{
Main MCMC sampler for Bayesian Distance Clustering with Pitman-Yor process priors.
Performs inference on cluster allocations, the mass parameter M, and the discount parameter theta.
}
\examples{
\dontrun{
# Set up parameters for a distance matrix D
params <- list(
  z.init = rep(1, nrow(D)),
  M.init = 10,
  theta.init = 0.5,
  metropolis_M_sd = 1,
  metropolis_theta_sd = 0.1,
  lambda.M = 1,
  alpha.theta = 1,
  beta.theta = 1,
  delta1 = 1, mu = 1, beta = 1,
  delta2 = 1, zeta = 1, gamma = 1
)
# Run sampler for 1000 iterations
samples <- mcmc_sampler(D, params, N.sim = 1000)
}

}
