% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tesselation.R
\name{mcmc_tesselation}
\alias{mcmc_tesselation}
\title{MCMC Sampler for Voronoi Tesselation with Birth-Death-Move}
\usage{
mcmc_tesselation(
  D,
  prior_param = 3,
  params,
  N.sim = 10000,
  tempering = 0,
  verbose = FALSE
)
}
\arguments{
\item{D}{Distance matrix (should be upper triangular with diagonal of zeros)}

\item{prior_param}{Prior parameter for number of clusters (default 3)}

\item{params}{List of model parameters including medoids.init}

\item{N.sim}{Number of MCMC iterations (default 10000)}

\item{tempering}{Tempering parameter (default 0)}

\item{verbose}{Logical; print iteration details (default FALSE)}
}
\value{
List of MCMC samples
}
\description{
Performs MCMC sampling over Voronoi tesselation configurations using birth-death-move updates.
}
