% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tesselation.R
\name{nested_tesselation}
\alias{nested_tesselation}
\title{Nested Voronoi Tesselation}
\usage{
nested_tesselation(D2, centers2, z1)
}
\arguments{
\item{D2}{Distance matrix for layer 2}

\item{centers2}{Center indices for layer 2}

\item{z1}{Clustering for layer 1 (constraint)}
}
\value{
Numeric vector of cluster assignments for z2
}
\description{
Computes a restricted Voronoi tesselation for z2 that respects the partition structure of z1.
}
\keyword{internal}
