% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tessellation_joint.R
\name{rand_index}
\alias{rand_index}
\title{Compute Rand Index Between Two Clusterings}
\usage{
rand_index(z1, z2)
}
\arguments{
\item{z1}{First clustering vector}

\item{z2}{Second clustering vector}
}
\value{
Numeric value of Rand index
}
\description{
Computes the Rand index agreement measure between two clustering solutions.
}
\keyword{internal}
