% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_predictive_sampler.R
\name{sample_from_pitman_yor_process}
\alias{sample_from_pitman_yor_process}
\title{Sample from Pitman-Yor Process with Gaussian base distribution}
\usage{
sample_from_pitman_yor_process(N, alpha, theta, n_approx = 10000)
}
\arguments{
\item{N}{Number of samples}

\item{alpha}{Mass parameter}

\item{theta}{Discount parameter}

\item{n_approx}{Number of approximation terms}
}
\value{
Numeric vector of cluster assignments
}
\description{
Generates samples from a Pitman-Yor process using stick-breaking construction.
}
