% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{telescopic_dependence}
\alias{telescopic_dependence}
\title{Telescopic Dependence Between Clusterings}
\usage{
telescopic_dependence(z1, z2, rcpp = TRUE)
}
\arguments{
\item{z1}{First clustering vector}

\item{z2}{Second clustering vector}

\item{rcpp}{Logical; use Rcpp implementation}
}
\value{
Numeric value measuring dependence
}
\description{
Measures dependence between two clustering solutions.
}
\keyword{internal}
