% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_tesselation.R
\name{tesselation}
\alias{tesselation}
\title{Compute Voronoi Tesselation}
\usage{
tesselation(D, centers)
}
\arguments{
\item{D}{Distance matrix}

\item{centers}{Vector of center indices}
}
\value{
Numeric vector of cluster assignments
}
\description{
Assigns each point to its nearest center (Voronoi assignment).
}
\keyword{internal}
