% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_matrix}
\alias{to_matrix}
\title{Convert Output List to Matrix}
\usage{
to_matrix(output, output_name)
}
\arguments{
\item{output}{List of outputs from MCMC}

\item{output_name}{Name of element to extract}
}
\value{
Matrix of extracted values
}
\description{
Extracts named elements from a list of outputs and converts to matrix.
}
