% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{update_gammas}
\alias{update_gammas}
\title{Update Gamma Indicators}
\usage{
update_gammas(z1, z2, M, theta, gamma.d, alpha.d, rcpp = TRUE)
}
\arguments{
\item{z1}{First clustering vector}

\item{z2}{Second clustering vector}

\item{M}{Mass parameter}

\item{theta}{Discount parameter}

\item{gamma.d}{Binary indicator vector}

\item{alpha.d}{Concentration parameter}

\item{rcpp}{Logical; use Rcpp implementation}
}
\value{
Updated gamma.d vector
}
\description{
Updates binary indicators that determine shared structure between two clusterings.
}
\keyword{internal}
