% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{update_M}
\alias{update_M}
\title{Update M Parameter in Joint MCMC Sampler}
\usage{
update_M(z1, z2, M, theta, gamma.d, params, rcpp)
}
\arguments{
\item{z1}{First clustering vector}

\item{z2}{Second clustering vector}

\item{M}{Current value of M parameter}

\item{theta}{Current value of theta parameter}

\item{gamma.d}{Binary indicator vector}

\item{params}{Model parameters including metropolis_M_sd and lambda.M}

\item{rcpp}{Logical; use Rcpp implementation}
}
\value{
Updated value of M parameter
}
\description{
Performs Metropolis-Hastings step to update the M parameter in joint sampler.
}
\keyword{internal}
