% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY_marginal.R
\name{update_M_marginal}
\alias{update_M_marginal}
\title{Update M Parameter in Marginal MCMC Sampler}
\usage{
update_M_marginal(z, M, theta, params, rcpp)
}
\arguments{
\item{z}{Numeric vector of cluster assignments}

\item{M}{Current value of M parameter}

\item{theta}{Current value of theta parameter}

\item{params}{List of model parameters including metropolis_M_sd and lambda.M}

\item{rcpp}{Logical; whether to use Rcpp implementations}
}
\value{
Updated value of M parameter
}
\description{
Performs a Metropolis-Hastings step to update the M parameter using truncated normal proposals.
}
\keyword{internal}
