% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{update_theta}
\alias{update_theta}
\title{Update Theta Parameter in Joint MCMC Sampler}
\usage{
update_theta(z1, z2, M, theta, gamma.d, params, rcpp)
}
\arguments{
\item{z1}{First clustering vector}

\item{z2}{Second clustering vector}

\item{M}{Current value of M parameter}

\item{theta}{Current value of theta parameter}

\item{gamma.d}{Binary indicator vector}

\item{params}{Model parameters including metropolis_theta_sd and beta prior parameters}

\item{rcpp}{Logical; use Rcpp implementation}
}
\value{
Updated value of theta parameter
}
\description{
Performs Metropolis-Hastings step to update the theta parameter in joint sampler.
}
\keyword{internal}
