% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY_marginal.R
\name{update_theta_marginal}
\alias{update_theta_marginal}
\title{Update Theta Parameter in Marginal MCMC Sampler}
\usage{
update_theta_marginal(z, M, theta, params, rcpp)
}
\arguments{
\item{z}{Numeric vector of cluster assignments}

\item{M}{Current value of M parameter}

\item{theta}{Current value of theta parameter}

\item{params}{List of model parameters}

\item{rcpp}{Logical; whether to use Rcpp implementations}
}
\value{
Updated value of theta parameter
}
\description{
Performs a Metropolis-Hastings step to update the theta parameter using truncated normal proposals.
}
\keyword{internal}
