% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{update_z}
\alias{update_z}
\title{Update Single Cluster Assignment}
\usage{
update_z(z, j, log_probs, rcpp = TRUE)
}
\arguments{
\item{z}{Clustering vector}

\item{j}{Index of element to update}

\item{log_probs}{Numeric vector of log probabilities for each cluster}

\item{rcpp}{Logical; use Rcpp implementation}
}
\value{
Updated clustering vector
}
\description{
Updates cluster assignment for element j using Gibbs sampling with cluster relabeling for singletons.
}
\keyword{internal}
