% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{update_z1}
\alias{update_z1}
\title{Update First Clustering Vector}
\usage{
update_z1(z1, z2, gamma.d, M, theta, D1, params, rcpp)
}
\arguments{
\item{z1}{First clustering vector}

\item{z2}{Second clustering vector}

\item{gamma.d}{Binary indicator vector}

\item{M}{Mass parameter}

\item{theta}{Discount parameter}

\item{D1}{First distance matrix}

\item{params}{Model parameters}

\item{rcpp}{Logical; use Rcpp implementation}
}
\value{
Updated z1 vector
}
\description{
Updates cluster allocations in z1 with compatibility constraints from z2 and gamma.
}
\keyword{internal}
