% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{update_z2}
\alias{update_z2}
\title{Update Second Clustering Vector}
\usage{
update_z2(z1, z2, gamma.d, M, theta, D2, params, rcpp)
}
\arguments{
\item{z1}{First clustering vector}

\item{z2}{Second clustering vector}

\item{gamma.d}{Binary indicator vector}

\item{M}{Mass parameter}

\item{theta}{Discount parameter}

\item{D2}{Second distance matrix}

\item{params}{Model parameters}

\item{rcpp}{Logical; use Rcpp implementation}
}
\value{
Updated z2 vector
}
\description{
Updates cluster allocations in z2, only for observations where gamma=0.
}
\keyword{internal}
