% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WSS_kmeans}
\alias{WSS_kmeans}
\title{Within-Cluster Sum of Squares for K-means}
\usage{
WSS_kmeans(D, Kmax)
}
\arguments{
\item{D}{Distance matrix}

\item{Kmax}{Maximum number of clusters to evaluate}
}
\description{
Plots WSS as function of number of clusters using k-means.
}
\keyword{internal}
