% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{WSS_kmedoids}
\alias{WSS_kmedoids}
\title{Within-Cluster Sum of Squares for K-medoids}
\usage{
WSS_kmedoids(D, Kmax)
}
\arguments{
\item{D}{Distance matrix}

\item{Kmax}{Maximum number of clusters to evaluate}
}
\description{
Plots WSS as function of number of clusters using fast k-medoids.
}
\keyword{internal}
