% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_EB_params}
\alias{compute_EB_params}
\title{Compute Empirical Bayes Parameters}
\usage{
compute_EB_params(D, k, k2 = NULL, linear = FALSE, repulsion = TRUE)
}
\arguments{
\item{D}{Distance matrix}

\item{k}{Number of clusters for initial k-medoids}

\item{k2}{Optional cluster assignments; if NULL, k-medoids clustering is performed}

\item{linear}{Logical; if TRUE uses linear model, if FALSE uses quadratic (default FALSE)}

\item{repulsion}{Optional parameter: will the sampler consider the repulsive term in the marginal likelihood? (default TRUE)}
}
\value{
List of hyperparameters for the model
}
\description{
Computes empirical Bayes hyperparameters from data using k-medoids clustering.
}
