% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_params.R
\name{likelihood_params}
\alias{likelihood_params}
\title{Likelihood Parameters}
\description{
A list specifying the parameters for the likelihood model.
}
\section{Quadratic Likelihood (`type = "quadratic"`)}{

For the quadratic likelihood, which is the default, the list should contain the following items:
\itemize{
  \item `type`: A string, must be `"quadratic"`.
  \item `shape_within`: Shape parameter for the Gamma distribution of within-cluster distances.
  \item `prior_shape_within`: Shape parameter for the Gamma prior on the rate of within-cluster distances.
  \item `prior_rate_within`: Rate parameter for the Gamma prior on the rate of within-cluster distances.
  \item `shape_between`: Shape parameter for the Gamma distribution of between-cluster distances.
  \item `prior_shape_between`: Shape parameter for the Gamma prior on the rate of between-cluster distances.
  \item `prior_rate_between`: Rate parameter for the Gamma prior on the rate of between-cluster distances.
  \item `repulsion`: A boolean indicating whether to model between-cluster repulsion.
}
}

\section{Linear Likelihood (`type = "linear"`)}{

For the linear likelihood, which models distances to and between cluster medoids, the list should contain:
\itemize{
  \item `type`: A string, must be `"linear"`.
  \item `shape_within`: Shape parameter for the Gamma distribution of within-cluster distances (to the medoid).
  \item `prior_shape_within`: Shape parameter for the Gamma prior on the rate of within-cluster distances.
  \item `prior_rate_within`: Rate parameter for the Gamma prior on the rate of within-cluster distances.
  \item `shape_between`: Shape parameter for the Gamma distribution of between-medoid distances.
  \item `rate_between`: Rate parameter for the Gamma distribution of between-medoid distances.
  \item `repulsion`: A boolean indicating whether to model between-cluster repulsion.
}
}

