% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mcmc_tessellation}
\alias{mcmc_tessellation}
\title{Run Reversible Jump MCMC for Voronoi Tessellation}
\usage{
mcmc_tessellation(
  distance_matrix,
  likelihood_params,
  prior_params,
  algo_params
)
}
\description{
Run Reversible Jump MCMC for Voronoi Tessellation
}
