% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_sampler_PY.R
\name{unlogprob_to_prob}
\alias{unlogprob_to_prob}
\title{Convert Log Probabilities to Normalized Probabilities}
\usage{
unlogprob_to_prob(log_prob)
}
\arguments{
\item{log_prob}{Numeric vector of log probabilities}
}
\value{
Numeric vector of normalized probabilities
}
\description{
Converts unormalized log probabilities to normalized probability vector.
}
\keyword{internal}
